+++ crates/config/src/schema/renderers/json_schema.rs
@@ -20,6 +20,17 @@ fn clean_comment(comment: String) -> String {
     comment.trim().replace('\n', " ")
 }
 
+fn lit_to_value(lit: &LiteralValue) -> Value {
+    match lit {
+        LiteralValue::Bool(inner) => Value::Bool(*inner),
+        LiteralValue::F32(inner) => Value::Number(Number::from_f64(*inner as f64).unwrap()),
+        LiteralValue::F64(inner) => Value::Number(Number::from_f64(*inner).unwrap()),
+        LiteralValue::Int(inner) => Value::Number(Number::from(*inner)),
+        LiteralValue::UInt(inner) => Value::Number(Number::from(*inner)),
+        LiteralValue::String(inner) => Value::String(inner.to_owned()),
+    }
+}
+
 impl JsonSchemaRenderer {
     pub fn new(options: JsonSchemaOptions) -> Self {
         Self {
@@ -32,14 +43,20 @@ impl JsonSchemaRenderer {
         let mut schema = self.render_schema(&field.type_of)?;
 
         if let Schema::Object(ref mut inner) = schema {
-            inner.metadata = Some(Box::new(Metadata {
+            let mut metadata = Metadata {
                 // title: field.name.clone(),
                 description: field.description.clone().map(clean_comment),
                 deprecated: field.deprecated,
                 read_only: field.read_only,
                 write_only: field.write_only,
                 ..Default::default()
-            }));
+            };
+
+            if let Some(default) = field.type_of.get_default() {
+                metadata.default = Some(lit_to_value(default));
+            }
+
+            inner.metadata = Some(Box::new(metadata));
         }
 
         Ok(schema)
@@ -80,7 +97,7 @@ impl SchemaRenderer<Schema> for JsonSchemaRenderer {
         Ok(Schema::Object(data))
     }
 
-    fn render_boolean(&mut self) -> RenderResult<Schema> {
+    fn render_boolean(&mut self, _boolean: &BooleanType) -> RenderResult<Schema> {
         Ok(Schema::Object(SchemaObject {
             instance_type: Some(SingleOrVec::Single(Box::new(InstanceType::Boolean))),
             ..Default::default()
@@ -175,17 +192,8 @@ impl SchemaRenderer<Schema> for JsonSchemaRenderer {
 
     fn render_literal(&mut self, literal: &LiteralType) -> RenderResult<Schema> {
         if let Some(value) = &literal.value {
-            let value = match value {
-                LiteralValue::Bool(inner) => Value::Bool(*inner),
-                LiteralValue::F32(inner) => Value::Number(Number::from_f64(*inner as f64).unwrap()),
-                LiteralValue::F64(inner) => Value::Number(Number::from_f64(*inner).unwrap()),
-                LiteralValue::Int(inner) => Value::Number(Number::from(*inner)),
-                LiteralValue::UInt(inner) => Value::Number(Number::from(*inner)),
-                LiteralValue::String(inner) => Value::String(inner.to_owned()),
-            };
-
             return Ok(Schema::Object(SchemaObject {
-                const_value: Some(value),
+                const_value: Some(lit_to_value(value)),
                 ..Default::default()
             }));
         }
@@ -304,8 +312,20 @@ impl SchemaRenderer<Schema> for JsonSchemaRenderer {
     fn render_union(&mut self, uni: &UnionType) -> RenderResult<Schema> {
         let mut items = vec![];
 
+        let mut metadata = Metadata {
+            title: uni.name.clone(),
+            description: uni.description.clone().map(clean_comment),
+            ..Default::default()
+        };
+
         for item in &uni.variants_types {
             items.push(self.render_schema(item)?);
+
+            if metadata.default.is_none() {
+                if let Some(def) = item.get_default() {
+                    metadata.default = Some(lit_to_value(def));
+                }
+            }
         }
 
         let subschema = match uni.operator {
@@ -320,6 +340,7 @@ impl SchemaRenderer<Schema> for JsonSchemaRenderer {
         };
 
         Ok(Schema::Object(SchemaObject {
+            metadata: Some(Box::new(metadata)),
             subschemas: Some(Box::new(subschema)),
             ..Default::default()
         }))