+++ crates/macros/src/config/setting.rs
@@ -5,7 +5,7 @@ use crate::utils::{
 use darling::FromAttributes;
 use proc_macro2::{Ident, TokenStream};
 use quote::{quote, ToTokens};
-use syn::{Attribute, Expr, ExprPath, Field, Type};
+use syn::{Attribute, Expr, ExprPath, Field, Lit, Type};
 
 // #[serde()]
 #[derive(FromAttributes, Default)]
@@ -263,11 +263,37 @@ impl<'l> Setting<'l> {
             }
         };
 
-        let type_of = if partial {
+        let mut type_of = if partial {
             quote! { SchemaType::infer_partial::<#value>() }
         } else {
             quote! { SchemaType::infer::<#value>() }
         };
+
+        if let Some(default) = &self.args.default {
+            if let Expr::Lit(lit) = &default {
+                let lit_value = match &lit.lit {
+                    Lit::Str(v) => quote! { LiteralValue::String(#v.into()) },
+                    Lit::Int(v) => {
+                        if v.suffix().starts_with('u') {
+                            quote! { LiteralValue::Uint(#v) }
+                        } else {
+                            quote! { LiteralValue::Int(#v) }
+                        }
+                    }
+                    Lit::Float(v) => {
+                        if v.suffix() == "f32" {
+                            quote! { LiteralValue::F32(#v) }
+                        } else {
+                            quote! { LiteralValue::F64(#v) }
+                        }
+                    }
+                    Lit::Bool(v) => quote! { LiteralValue::Bool(#v) },
+                    _ => unimplemented!(),
+                };
+
+                type_of = quote! { SchemaType::infer_with_default::<#value>(#lit_value) };
+            }
+        }
 
         quote! {
             SchemaField {