+++ crates/types/src/lib.rs
@@ -1,4 +1,5 @@
 mod arrays;
+mod bools;
 mod enums;
 mod externals;
 mod literals;
@@ -10,6 +11,7 @@ mod tuples;
 mod unions;
 
 pub use arrays::*;
+pub use bools::*;
 pub use enums::*;
 pub use externals::*;
 pub use literals::*;
@@ -23,11 +25,11 @@ pub use unions::*;
 /// All possible types within a schema.
 #[derive(Clone, Debug, Default)]
 pub enum SchemaType {
-    Boolean,
     Null,
     #[default]
     Unknown,
     Array(ArrayType),
+    Boolean(BooleanType),
     Enum(EnumType),
     Float(FloatType),
     Integer(IntegerType),
@@ -46,6 +48,14 @@ impl SchemaType {
     }
 
     /// Infer a schema from a type that implements [`Schematic`],
+    /// and also provide a default literal value.
+    pub fn infer_with_default<T: Schematic>(default: LiteralValue) -> SchemaType {
+        let mut schema = T::generate_schema();
+        schema.set_default(default);
+        schema
+    }
+
+    /// Infer a schema from a type that implements [`Schematic`],
     /// and mark the schema is partial (is marked as `nested`).
     pub fn infer_partial<T: Schematic>() -> SchemaType {
         let mut schema = T::generate_schema();
@@ -59,6 +69,11 @@ impl SchemaType {
             items_type: Box::new(items_type),
             ..ArrayType::default()
         })
+    }
+
+    /// Create a boolean type.
+    pub fn boolean() -> SchemaType {
+        SchemaType::Boolean(BooleanType::default())
     }
 
     /// Create a float schema with the provided kind.
@@ -167,13 +182,33 @@ impl SchemaType {
         })
     }
 
+    /// Return a `default` value from the inner schema type.
+    pub fn get_default(&self) -> Option<&LiteralValue> {
+        match self {
+            SchemaType::Boolean(BooleanType { default, .. }) => default.as_ref(),
+            SchemaType::Float(FloatType { default, .. }) => default.as_ref(),
+            SchemaType::Integer(IntegerType { default, .. }) => default.as_ref(),
+            SchemaType::String(StringType { default, .. }) => default.as_ref(),
+            SchemaType::Union(UnionType { variants_types, .. }) => {
+                for variant in variants_types {
+                    if let Some(value) = variant.get_default() {
+                        return Some(value);
+                    }
+                }
+
+                None
+            }
+            _ => None,
+        }
+    }
+
     /// Return a `name` from the inner schema type.
     pub fn get_name(&self) -> Option<&String> {
         match self {
-            SchemaType::Boolean => None,
             SchemaType::Null => None,
             SchemaType::Unknown => None,
             SchemaType::Array(ArrayType { name, .. }) => name.as_ref(),
+            SchemaType::Boolean(BooleanType { name, .. }) => name.as_ref(),
             SchemaType::Enum(EnumType { name, .. }) => name.as_ref(),
             SchemaType::Float(FloatType { name, .. }) => name.as_ref(),
             SchemaType::Integer(IntegerType { name, .. }) => name.as_ref(),
@@ -184,6 +219,25 @@ impl SchemaType {
             SchemaType::Tuple(TupleType { name, .. }) => name.as_ref(),
             SchemaType::Union(UnionType { name, .. }) => name.as_ref(),
         }
+    }
+
+    /// Set the `default` of the inner schema type.
+    pub fn set_default(&mut self, default: LiteralValue) {
+        match self {
+            SchemaType::Boolean(ref mut inner) => {
+                inner.default = Some(default);
+            }
+            SchemaType::Float(ref mut inner) => {
+                inner.default = Some(default);
+            }
+            SchemaType::Integer(ref mut inner) => {
+                inner.default = Some(default);
+            }
+            SchemaType::String(ref mut inner) => {
+                inner.default = Some(default);
+            }
+            _ => {}
+        };
     }
 
     /// Set the `name` of the inner schema type. If the inner type does not support